+++ terra-rust-api/Changelog.md
@@ -2,6 +2,7 @@
 
 ## 0.2
 ### 0.2.8 - devel
+* Add slashing/Unjail
 ### 0.2.7 - 12-Aug-21
 * Add timezone support (used on tendermint dates in addressbook)
 * Add functions to fetch address book